refactor : 다형성을 활용한 리팩토링
