+++ src/main/java/wooteco/subway/member/application/MemberService.java
@@ -4,8 +4,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import wooteco.subway.exception.conflict.AlreadyExistEmailException;
 import wooteco.subway.member.dao.MemberDao;
-import wooteco.subway.member.domain.LoginMember;
 import wooteco.subway.member.domain.Member;
+import wooteco.subway.member.domain.authmember.LoginMember;
 import wooteco.subway.member.dto.MemberEmailCheckRequest;
 import wooteco.subway.member.dto.MemberRequest;
 import wooteco.subway.member.dto.MemberResponse;
