+++ src/main/java/wooteco/subway/member/domain/agemember/AgeMember.java
@@ -1,0 +1,8 @@
+package wooteco.subway.member.domain.agemember;
+
+public interface AgeMember {
+
+    boolean isInAgeRange(int age);
+
+    int discountFareByAge(int fare);
+}
