+++ src/main/java/wooteco/subway/member/domain/authmember/AuthMember.java
@@ -1,6 +1,8 @@
-package wooteco.subway.member.domain;
+package wooteco.subway.member.domain.authmember;
 
 public interface AuthMember {
+
     int discountFareByAge(int fare);
+
     boolean isLoggedIn();
 }
