+++ src/main/java/wooteco/subway/member/ui/MemberController.java
@@ -15,8 +15,8 @@ import wooteco.subway.auth.domain.AuthenticationPrincipal;
 import wooteco.subway.exception.unauthorized.AuthorizationException;
 import wooteco.subway.exception.unauthorized.NotLoginException;
 import wooteco.subway.member.application.MemberService;
-import wooteco.subway.member.domain.AuthMember;
-import wooteco.subway.member.domain.LoginMember;
+import wooteco.subway.member.domain.authmember.AuthMember;
+import wooteco.subway.member.domain.authmember.LoginMember;
 import wooteco.subway.member.dto.MemberEmailCheckRequest;
 import wooteco.subway.member.dto.MemberRequest;
 import wooteco.subway.member.dto.MemberResponse;
