+++ src/main/java/wooteco/subway/path/dto/PathResponseAssembler.java
@@ -2,7 +2,7 @@ package wooteco.subway.path.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import wooteco.subway.member.domain.AuthMember;
+import wooteco.subway.member.domain.authmember.AuthMember;
 import wooteco.subway.path.domain.SubwayPath;
 import wooteco.subway.station.dto.StationResponse;
